pyproject.toml
[project]
name = "dynamic-leadgen"
version = "0.1.0"
description = "Dynamic Lead Gen with Agents PoC"
authors = [
    { name = "Julie Berlin", email = "ai@jkwb.mozmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.5.4",
    "langgraph-sdk>=0.1.74",
    "openai>=1.97.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.17.0",
    "uvicorn>=0.35.0",
]
